% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{argument_match_test}
\alias{argument_match_test}
\title{Test function with non-matching arguments}
\usage{
argument_match_test(fun, args)
}
\arguments{
\item{fun}{Function to test}

\item{args}{Complete \code{list} of required arguments to \code{fun}}
}
\description{
Goes through all the parameters of fun which have a vector of default values. It then calls \code{fun} with a different value.
}

