% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{capitalize_first}
\alias{capitalize_first}
\title{Capitalize the first letter of a word}
\usage{
capitalize_first(word)
}
\arguments{
\item{word}{A string}
}
\value{
String with first letter converted to capital
}
\description{
Takes a string, and converts its first letter to upper case
}

