% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multitest.R
\name{create_and_run_test}
\alias{create_and_run_test}
\title{Create test and run it}
\usage{
create_and_run_test(train_index, data, dependent,
  problem = c("classification", "regression"), method = c("randomForest",
  "rpart"), name, description, data_transform = quote(identity))
}
\arguments{
\item{train_index}{A vector of the rows to be used as training set. All other rows will form the holdout set}

\item{dependent}{The dependent variable: the name of the column containing the prediction goal}

\item{problem}{Either classification or regression. This influences how the algorithms are trained}

\item{method}{The regression or classification method}

\item{description}{Optional. A more elaborate description of the test}

\item{data_transform}{A quoted function name that transforms the data.}
}
\description{
A convenience function calling \code{\link{createtest}} first, then runs the test using \code{\link{runtest}}.
}

