% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crtests.R
\docType{package}
\name{crtests}
\alias{crtests}
\alias{crtests-package}
\title{crtests: A package for creating and executing classification and regression tests}
\description{
The crtests package provides functions to prepare data, train models, test models, and evaluate outcomes. Its goal is to provide a generic and extendable API to machine learning and statistics functions
}

