% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{drop_na}
\alias{drop_na}
\title{Remove NAs according to a strategy}
\usage{
drop_na(strategy = c("dependent", "predictors", "all", "none"), df, dependent)
}
\arguments{
\item{strategy}{Character string denoting how NAs should be dealt with. "dependent" means rows with NA in the dependent variable are dropped. "predictors" means rows with NA in an independent variable are dropped. "all" means rows with NA in any column are dropped. "none" means NAs are ignored.}

\item{df}{Data frame to remove NAs from}

\item{dependent}{Dependent variable of the data frame}
}
\value{
A data.frame where \code{strategy} has been applied to remove data
}
\description{
Remove NAs according to a strategy
}

