% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate the performance of a prediction.}
\usage{
evaluate(prediction, data, test, ...)
}
\arguments{
\item{prediction}{A vector of predictions for each row in the holdout set}

\item{data}{The data list containing train and holdout data sets}

\item{test}{The test object being evaluated}
}
\value{
A vector of predictions
}
\description{
Wraps the problem-specific evaluation functions by calling \code{\link{evaluate_problem}}. This wrapper is desirable, as it can perform the extraction of the holdout set (observations)
}

