% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_problem}
\alias{evaluate_problem}
\alias{evaluate_problem.classification}
\alias{evaluate_problem.regression}
\title{Generic function for evaluation of test results}
\usage{
evaluate_problem(test, prediction, observations)

\method{evaluate_problem}{classification}(test, prediction, observations)

\method{evaluate_problem}{regression}(test, prediction, observations)
}
\arguments{
\item{test}{The test that was run}

\item{observations}{The true observations for the dependent value in the holdout set}

\item{predictions}{A vector of predictions for each row in the holdout set}
}
\description{
Generic function for evaluation of test results
}
\section{Methods (by class)}{
\itemize{
\item \code{classification}: Evaluate a classification test's results. Calls \code{\link[caret]{confusionMatrix}}

\item \code{regression}: Evaluate a regression test's results
}}

