% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/train_helpers.R
\name{extract_formula}
\alias{extract_formula}
\title{Extract a formula from a test}
\usage{
extract_formula(test)
}
\arguments{
\item{test}{An object of class 'regression' or 'classification'}
}
\description{
Extracts a formula of the form \code{dependent ~ .} from the test object
}

