% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{factor_length}
\alias{factor_length}
\title{Determine the length of the factors in a data.frame}
\usage{
factor_length(df)
}
\arguments{
\item{df}{A data.frame}
}
\value{
A vector of length \code{n}, with \code{n} the number of factor columns in the data frame, containing the length of the levels of those factors
}
\description{
Goes through every column in the data.frame, and return the length of its levels
}

