% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{group_levels.default}
\alias{group_levels.default}
\title{Group infrequent factor levels}
\usage{
\method{group_levels}{default}(data, maximum_levels = 32)
}
\arguments{
\item{data}{A data.frame or factor. In the first case, \code{group_levels} is applied to each factor in the data.frame.}

\item{maximum_levels}{Numeric. The maximum number of levels allowed per factor}
}
\description{
The default group_levels does nothing. This is desirable behavior for any structure that is not a list, data.frame or factor: there is no meaningful way apply group_levels to this type of structure.
}

