% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{is_complete_row}
\alias{is_complete_row}
\title{Determine if the rows in a data.frame have NAs}
\usage{
is_complete_row(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A vector of length \code{nrow(data)} containing whether that row has \code{\link{NA}}s.
}
\description{
Determine if the rows in a data.frame have NAs
}

