% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_prepare.R
\name{method_prepare.rpart}
\alias{method_prepare.rpart}
\title{#'@describeIn method_prepare RPart specific data preparation. Identical to \code{\link{method_prepare.randomForest}}}
\usage{
\method{method_prepare}{rpart}(method, test, ...)
}
\description{
#'@describeIn method_prepare RPart specific data preparation. Identical to \code{\link{method_prepare.randomForest}}
}

