% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.default}
\alias{prepare.default}
\title{Prepare the data for the specified test.
The default method relevels the holdout set, so the holdout and train set are completely independent, and to prevent problems with certain algorithms that can't deal with different factor levels across train and holdout set}
\usage{
\method{prepare}{default}(test, ...)
}
\description{
Prepare the data for the specified test.
The default method relevels the holdout set, so the holdout and train set are completely independent, and to prevent problems with certain algorithms that can't deal with different factor levels across train and holdout set
}

