% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitest_helpers.R
\name{print.multitest_evaluation}
\alias{print.multitest_evaluation}
\title{Print a multitest_evaluation}
\usage{
\method{print}{multitest_evaluation}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a multitest_evaluation
}

