% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{random_string}
\alias{random_string}
\title{Generate a random string}
\usage{
random_string(length)
}
\arguments{
\item{length}{Length of string to generate}
}
\value{
If \code{length>0}: A random sequence of characters of length \code{length}, otherwise an empty string
}
\description{
Generates a random string of length \code{length}
}

