% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_names.matrix}
\alias{remove_names.matrix}
\title{Set all row and column names of a matrix to ""}
\usage{
\method{remove_names}{matrix}(x)
}
\arguments{
\item{matrix}{Matrix to 'remove' colnames and rownames from}
}
\value{
Matrix where colnames and rownames consist of only ""
}
\description{
Set row and column names to "" for pretty printing
}

