% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_names}
\alias{remove_names}
\title{Set any names of x to ""}
\usage{
remove_names(x)
}
\arguments{
\item{x}{An object that has a 'names' property, typically a matrix, list or data.frame}
}
\description{
Set any names of x to ""
}

