% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replace_names.default}
\alias{replace_names.default}
\title{Default method that replaces names(object)}
\usage{
\method{replace_names}{default}(object, pattern = "\\\\.",
  replacement = " ", ...)
}
\arguments{
\item{object}{Object of which the names are to be changed}

\item{pattern}{Pattern to look for, as defined by a POSIX regular expression}

\item{replacement}{Replacement string}

\item{...}{extra arguments to \code{replace_names}}
}
\description{
Default method that replaces names(object)
}

