% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replace_names.matrix}
\alias{replace_names.matrix}
\title{Replace row.names and col.names in the object}
\usage{
\method{replace_names}{matrix}(object, pattern = "\\\\.", replacement = " ",
  replace_rownames = TRUE, replace_colnames = TRUE, ...)
}
\arguments{
\item{object}{Object of which the names are to be changed}

\item{pattern}{Pattern to look for, as defined by a POSIX regular expression}

\item{replacement}{Replacement string}

\item{replace_rownames}{Logical. Should row names be replaced?}

\item{replace_colnames}{Logical. Should column names be replaced?}

\item{...}{extra arguments to \code{replace_names}}
}
\description{
Replace row.names and col.names in the object
}

