% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replace_names}
\alias{replace_names}
\title{Replace strings in the names of an object}
\usage{
replace_names(object, pattern, replacement, ...)
}
\arguments{
\item{object}{Object of which the names are to be changed}

\item{pattern}{Pattern to look for, as defined by a POSIX regular expression}

\item{replacement}{Replacement string}

\item{...}{extra arguments to \code{replace_names}}
}
\description{
Replaces strings matching the pattern in the names of the object by the replacement. If applicable, both row and column names could be replaced. This function is a simple wrapper to \code{\link[stringr]{str_replace_all}}
}
\seealso{
\code{\link[stringr]{str_replace_all}}
}

