% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.multitest_results_classification}
\alias{summary.multitest_results_classification}
\title{Make a summary of multiple executions of a test}
\usage{
\method{summary}{multitest_results_classification}(object, ...)
}
\arguments{
\item{results_list}{A list of results of the test}
}
\value{
list of minimum accuracy, maximum accuracy, mean accuracy, standard deviation of accuracy
}
\description{
Summary implementation for the results of a multitest of a classification problem
}

