% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\alias{train_model.classification}
\alias{train_model.regression}
\title{Train a classification or regression model}
\usage{
train_model(test, data, ...)

\method{train_model}{classification}(test, data, ...)

\method{train_model}{regression}(test, data, ...)
}
\arguments{
\item{test}{The test object. This is passed so the method can be extracted.}

\item{data}{An object of class "regression" or "classification" with at least x, y, train and holdout}

\item{...}{Extra arguments to pass to the classification or regression method}
}
\description{
Generic function for training a model.
}
\section{Methods (by class)}{
\itemize{
\item \code{classification}: Train a model for classification using a classifier algorithm. This function wraps the actual classifier

\item \code{regression}: Train (fit) a regression model. This function wraps a regression algorithm.
}}

