% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare the data for the specified test.}
\usage{
prepare(test, ...)
}
\arguments{
\item{test}{The test for which data is prepared}
}
\value{
data	A list containing prepared train (\code{data$train}) and holdout (\code{data$holdout}) data frames. Extra method specific preparation is executed through a call to method_prepare
}
\description{
This allows for different implementations for regression or classification
}

