% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for training or testing.}
\usage{
prepare_data(df, df_reference, relevel = TRUE)
}
\arguments{
\item{df}{The data frame that is to be prepared}

\item{df_reference}{An optional reference data frame, whose factor levels are to be applied to df}

\item{relevel}{Logical. Should the df be releveled with df_reference's factor levels?}
}
\value{
A data frame stripped of missing values
}
\description{
This function removes all missing values, including those introduced after (optional) releveling
}

