% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regression_model.R
\name{regression_model}
\alias{regression_model}
\alias{regression_model.default}
\title{Fit a regression model}
\usage{
regression_model(method, formula, train_data, ...)

\method{regression_model}{default}(method, formula, training_data, ...)
}
\arguments{
\item{method}{The regression method to use}

\item{formula}{An object of class 'formula', used to fit a model to the data}

\item{data}{Train data used to fit the model}
}
\value{
model	The fitted model
}
\description{
Generic function for fitting a regression model
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default function for fitting a regression model

This \code{\link[base]{get}}s the method and calls it using formula and data
}}

