% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runtest.R
\name{runtest}
\alias{runtest}
\alias{runtest.default}
\title{Run a classification or regression test}
\usage{
runtest(test, ...)

\method{runtest}{default}(test, ...)
}
\arguments{
\item{test}{An object of class 'classification' or 'regression'}

\item{...}{Extra arguments to runtest}
}
\description{
Run a classification or regression test
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: The default test run subsequently calls \code{\link{prepare}}, \code{\link{train_model}}, \code{\link{make_predictions}},  \code{\link{evaluate}}
}}

