% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate_helpers.R
\name{summary.evaluation}
\alias{summary.evaluation}
\title{Summary of an evaluation}
\usage{
\method{summary}{evaluation}(object, include_test_attributes = TRUE, ...)
}
\arguments{
\item{object}{Evaluation object to make summary of}

\item{include_test_attributes}{Logical. Should all attributes of the test be included in the output?}

\item{...}{Extra arguments to \code{summary.evaluation}}
}
\description{
Produces a summary of an evaluation, consisting of the test attributes and the performance measures
}

