% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multitest_helpers.R
\name{summary.multitest_evaluation}
\alias{summary.multitest_evaluation}
\title{Make a summary of multiple test evaluations}
\usage{
\method{summary}{multitest_evaluation}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Object of class 'summary.multitest_evaluation'. Attributes are a list of \code{evaluation} objects,
}
\description{
Summary implementation for the results of a multitest
}

