% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util}
\alias{missing_argument_test}
\alias{util}
\title{Utility functions}
\usage{
missing_argument_test(fun, args, testthat = testthat::expect_error, outcomes)
}
\arguments{
\item{fun}{Function to test}

\item{args}{Complete \code{list} of required arguments to \code{fun}}

\item{testthat}{Test function to run on every iteration}

\item{outcomes}{List of length \code{args} with expected outcomes for each test. Names should match those of \code{args}}
}
\description{
Utility functions

Utility for testing how a function deals with missing required arguments. It calls the function \code{length(args)} times, each time
omitting one argument
}

